+++ mdadash/backend/analyses/acf.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 import MDAnalysis as mda
 import numpy as np
 from joblib import delayed
+from matplotlib.backends.backend_agg import FigureCanvasAgg
 from matplotlib.collections import LineCollection
 from scipy import integrate
 
@@ -227,6 +228,7 @@ class ACFAnalysis(WidgetBase):
     def _setup_plot(self):
         """Setup matplotlib plot"""
         self.fig, self.ax = plt.subplots()
+        self.canvas = FigureCanvasAgg(self.fig)
         (self.plot,) = self.ax.plot([], [], color="red", zorder=2)
         self.lc = LineCollection([], colors="gray", alpha=0.2, lw=0.5, zorder=1)
         self.ax.add_collection(self.lc)
@@ -305,7 +307,7 @@ class ACFAnalysis(WidgetBase):
         self.lc.set_segments(y2 if y2 is not None else [])
         self.ax.relim()
         self.ax.autoscale_view()
-        self.display_fig(self.fig)
+        self.display_canvas(self.canvas)
 
     def run_every_frame(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.run_every_frame` handler"""