+++ mdadash/backend/analyses/helix_analysis.py
@@ -9,6 +9,7 @@ from typing import ClassVar
 
 import matplotlib.pyplot as plt
 from joblib import delayed
+from matplotlib.backends.backend_agg import FigureCanvasAgg
 from MDAnalysis.analysis.helix_analysis import HELANAL
 
 from mdadash.backend.widgets.base import WidgetBase
@@ -181,6 +182,7 @@ class HelixAnalysis(WidgetBase):
     def _setup_plot(self):
         """Setup matplotlib plot"""
         self.fig, self.ax = plt.subplots()
+        self.canvas = FigureCanvasAgg(self.fig)
         (self.plot,) = self.ax.plot([], [])
         self.ax.grid(True)
         self._set_title()
@@ -284,7 +286,7 @@ class HelixAnalysis(WidgetBase):
         self.plot.set_data(self.x_values, self.y_values)
         self.ax.relim()
         self.ax.autoscale_view()
-        self.display_fig(self.fig)
+        self.display_canvas(self.canvas)
 
     def run_every_frame(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.run_every_frame` handler"""