+++ mdadash/backend/analyses/native_contacts.py
@@ -8,6 +8,7 @@ from typing import ClassVar
 
 import matplotlib.pyplot as plt
 from joblib import delayed
+from matplotlib.backends.backend_agg import FigureCanvasAgg
 from MDAnalysis.analysis import contacts
 
 from mdadash.backend.widgets.base import WidgetBase
@@ -204,6 +205,7 @@ class NativeContacts(WidgetBase):
     def _setup_plot(self):
         """Setup matplotlib plot"""
         self.fig, self.ax = plt.subplots()
+        self.canvas = FigureCanvasAgg(self.fig)
         (self.plot,) = self.ax.plot([], [])
         self.ax.set_ylabel("Fraction of contacts")
         self.ax.grid(True)
@@ -321,7 +323,7 @@ class NativeContacts(WidgetBase):
         self.plot.set_data(self.x_values, self.y_values)
         self.ax.relim()
         self.ax.autoscale_view()
-        self.display_fig(self.fig)
+        self.display_canvas(self.canvas)
 
     def run_every_frame(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.run_every_frame` handler"""